Refactor benchmark graph display into graphdisplay.js
