+++ asv/www/index.html
@@ -54,6 +54,9 @@
             src="summary.js">
     </script>
     <script language="javascript" type="text/javascript"
+            src="graphdisplay.js">
+    </script>
+    <script language="javascript" type="text/javascript"
             src="regressions.js">
     </script>
     <link href="asv.css" rel="stylesheet" type="text/css"/>
